All delta is double

+++ headers/animation.h
@@ -16,8 +16,8 @@ typedef struct GbAnimation {
     uint16_t dx;
     uint16_t dy;
 
-    uint32_t interval;
-    uint32_t accumulator;
+    double interval;
+    double accumulator;
 
     uint16_t frame_count;
     uint16_t current_frame;
@@ -32,10 +32,10 @@ GbAnimation *gb_anim_new_animation(
                                    uint16_t init_y,
                                    uint16_t dx,
                                    uint16_t dy,
-                                   uint16_t interval,
+                                   double interval,
                                    uint16_t frame_count,
                                    int8_t direction,
                                    GB_ANIM_TYPE type);
 
-uint8_t gb_anim_apply(SDL_Rect *src, uint32_t delta, GbAnimation *animation);
+uint8_t gb_anim_apply(SDL_Rect *src, double delta, GbAnimation *animation);
 #endif // ANIMATION_H_INCLUDED
